+++ Controller/netflixController.js
@@ -98,15 +98,29 @@ exports.updateNetflixAccount = async (req, res, next) => {
 exports.deleteNetflixAccount = (req, res, next) => {
   try {
     const { id } = req.params;
-  const sql = `DELETE FROM netflix WHERE id = ${id}`;
+    const sql = `SELECT * FROM netflix WHERE id = ${id}`;
 
-  connection.query(sql, error => {
-    if (error) throw error;
-    res.send('Delete Netflix Account');
-  });
+    connection.query(sql, (error, results) => {
+      if (error) throw error;
+
+      if (results.length === 0) {
+        res.status(404).send('netflix Account not found');
+      } else {
+        const account = results[0];
+        if (account.usado !== '0') {
+          res.status(405).send('Cannot delete netflix Account. Account is unused.');
+        } else {
+          const deleteSql = `DELETE FROM netflix WHERE id = ${id}`;
+          connection.query(deleteSql, deleteError => {
+            if (deleteError) throw deleteError;
+            res.send('Delete netflix Account');
+          });
+        }
+      }
+    });
   } catch (error) {
-    res.send('Error: ' + JSON.stringify(error));
-    console.error(error)
+    res.status(500).send('Error: ' + JSON.stringify(error));
+    console.error(error);
     next(error);
   }
 }
