+++ Controller/startPlusController.js
@@ -82,14 +82,28 @@ exports.updateStartPlusAccount = async (req, res, next) => {
 exports.deleteStartPlusAccount = (req, res, next) => {
   try {
     const { id } = req.params;
-    const sql = `DELETE FROM startplus WHERE id = ${id}`;
+    const sql = `SELECT * FROM startplus WHERE id = ${id}`;
 
-    connection.query(sql, (error) => {
+    connection.query(sql, (error, results) => {
       if (error) throw error;
-      res.send('Delete StartPlus Account');
+
+      if (results.length === 0) {
+        res.status(404).send('startplus Account not found');
+      } else {
+        const account = results[0];
+        if (account.usado !== '0') {
+          res.status(405).send('Cannot delete startplus Account. Account is unused.');
+        } else {
+          const deleteSql = `DELETE FROM startplus WHERE id = ${id}`;
+          connection.query(deleteSql, deleteError => {
+            if (deleteError) throw deleteError;
+            res.send('Delete startplus Account');
+          });
+        }
+      }
     });
   } catch (error) {
-    res.send('Error: ' + JSON.stringify(error));
+    res.status(500).send('Error: ' + JSON.stringify(error));
     console.error(error);
     next(error);
   }
