+++ Controller/youtubeController.js
@@ -15,7 +15,7 @@ exports.getAllYoutubeAccounts = (req, res, next) => {
 
 exports.getOneYoutubeAccount = (req, res, next) => {
   const { id } = req.params;
-  const sql = `SELECT * FROM youtube WHERE id = ${id} cod_usuario ORDER BY id DESC`;
+  const sql = `SELECT * FROM youtube WHERE id = ${id} ORDER BY id DESC`;
   connection.query(sql, (error, result) => {
     if (error) throw error;
 
@@ -82,14 +82,28 @@ exports.updateYoutubeAccount = async (req, res, next) => {
 exports.deleteYoutubeAccount = (req, res, next) => {
   try {
     const { id } = req.params;
-    const sql = `DELETE FROM youtube WHERE id = ${id}`;
+    const sql = `SELECT * FROM youtube WHERE id = ${id}`;
 
-    connection.query(sql, (error) => {
+    connection.query(sql, (error, results) => {
       if (error) throw error;
-      res.send('Delete Youtube Account');
+
+      if (results.length === 0) {
+        res.status(404).send('youtube Account not found');
+      } else {
+        const account = results[0];
+        if (account.usado !== '0') {
+          res.status(405).send('Cannot delete youtube Account. Account is unused.');
+        } else {
+          const deleteSql = `DELETE FROM youtube WHERE id = ${id}`;
+          connection.query(deleteSql, deleteError => {
+            if (deleteError) throw deleteError;
+            res.send('Delete youtube Account');
+          });
+        }
+      }
     });
   } catch (error) {
-    res.send('Error: ' + JSON.stringify(error));
+    res.status(500).send('Error: ' + JSON.stringify(error));
     console.error(error);
     next(error);
   }
